requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
Django==4.2
django-crispy-forms==2.0
django-etc==1.4.0
django-hitcount==1.3.5
django-jazzmin==2.6.0
django-modeltranslation==0.18.9
Pillow==9.5.0
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.5.0
